requirements.txt
pydantic
pydantic-ai
python-dotenv
rich